Isolated the Draft class to avoid loop import
